# Optimize Operator Input Handling