--- a/go/vt/vtgate/planbuilder/operators/comments.go
+++ b/go/vt/vtgate/planbuilder/operators/comments.go
@@ -26,23 +26,23 @@ import (
 
 // LockAndComment contains any comments or locking directives we want on all queries down from this operator
 type LockAndComment struct {
-	Source   Operator
+	SingleSource
 	Comments *sqlparser.ParsedComments
 	Lock     sqlparser.Lock
+}
+
+func newLockAndComment(op Operator, comments *sqlparser.ParsedComments, lock sqlparser.Lock) Operator {
+	return &LockAndComment{
+		SingleSource: SingleSource{Source: op},
+		Comments:     comments,
+		Lock:         lock,
+	}
 }
 
 func (l *LockAndComment) Clone(inputs []Operator) Operator {
 	klon := *l
 	klon.Source = inputs[0]
 	return &klon
-}
-
-func (l *LockAndComment) Inputs() []Operator {
-	return []Operator{l.Source}
-}
-
-func (l *LockAndComment) SetInputs(operators []Operator) {
-	l.Source = operators[0]
 }
 
 func (l *LockAndComment) AddPredicate(ctx *plancontext.PlanningContext, expr sqlparser.Expr) Operator {