--- a/go/vt/vtgate/planbuilder/operators/filter.go
+++ b/go/vt/vtgate/planbuilder/operators/filter.go
@@ -45,7 +45,7 @@ func newFilterSinglePredicate(op Operator, expr sqlparser.Expr) Operator {
 
 func newFilter(op Operator, expr ...sqlparser.Expr) Operator {
 	return &Filter{
-		unaryOperator: unaryOperator{Source: op},
+		unaryOperator: newUnaryOp(op),
 		Predicates:    expr,
 	}
 }