--- a/go/vt/vtgate/planbuilder/operators/horizon.go
+++ b/go/vt/vtgate/planbuilder/operators/horizon.go
@@ -35,7 +35,7 @@ import (
 // Project/Aggregate/Sort/Limit operations, some which can be pushed down,
 // and some that have to be evaluated at the vtgate level.
 type Horizon struct {
-	Source Operator
+	SingleSource
 
 	// If this is a derived table, the two following fields will contain the tableID and name of it
 	TableId       *semantics.TableSet
@@ -55,7 +55,10 @@ type Horizon struct {
 }
 
 func newHorizon(src Operator, query sqlparser.SelectStatement) *Horizon {
-	return &Horizon{Source: src, Query: query}
+	return &Horizon{
+		SingleSource: SingleSource{Source: src},
+		Query:        query,
+	}
 }
 
 // Clone implements the Operator interface
@@ -76,16 +79,6 @@ func (h *Horizon) Clone(inputs []Operator) Operator {
 // if they do some things, like LIMIT or GROUP BY on wrong columns
 func (h *Horizon) IsMergeable(ctx *plancontext.PlanningContext) bool {
 	return isMergeable(ctx, h.Query, h)
-}
-
-// Inputs implements the Operator interface
-func (h *Horizon) Inputs() []Operator {
-	return []Operator{h.Source}
-}
-
-// SetInputs implements the Operator interface
-func (h *Horizon) SetInputs(ops []Operator) {
-	h.Source = ops[0]
 }
 
 func (h *Horizon) AddPredicate(ctx *plancontext.PlanningContext, expr sqlparser.Expr) Operator {