… go/vt/vtgate/planbuilder/operators/hash_join_test.go
… go/vt/vtgate/planbuilder/operators/insert.go
… go/vt/vtgate/planbuilder/operators/insert_selection.go
… go/vt/vtgate/planbuilder/operators/join.go
… go/vt/vtgate/planbuilder/operators/join_merging.go
… go/vt/vtgate/planbuilder/operators/limit.go
… go/vt/vtgate/planbuilder/operators/mirror.go
… go/vt/vtgate/planbuilder/operators/offset_planning.go
… go/vt/vtgate/planbuilder/operators/operator.go
… go/vt/vtgate/planbuilder/operators/ordering.go
… go/vt/vtgate/planbuilder/operators/phases.go
… go/vt/vtgate/planbuilder/operators/plan_query.go
… go/vt/vtgate/planbuilder/operators/projection.go
… go/vt/vtgate/planbuilder/operators/query_planning.go
… go/vt/vtgate/planbuilder/operators/querygraph.go
… go/vt/vtgate/planbuilder/operators/recurse_cte.go
… go/vt/vtgate/planbuilder/operators/rewriters.go
… go/vt/vtgate/planbuilder/operators/route.go
… go/vt/vtgate/planbuilder/operators/route_planning.go
… go/vt/vtgate/planbuilder/operators/subquery.go
… go/vt/vtgate/planbuilder/operators/subquery_planning.go
… go/vt/vtgate/planbuilder/operators/table.go
… go/vt/vtgate/planbuilder/operators/union_merging.go
… go/vt/vtgate/planbuilder/operators/update.go
… go/vt/vtgate/planbuilder/operators/vindex.go
… go/vt/vtgate/planbuilder/plan_test.go
… go/vt/vtgate/planbuilder/plancontext/planning_context.go
… go/vt/vtgate/planbuilder/testdata/filter_cases.json
… go/vt/vtgate/planbuilder/testdata/foreignkey_cases.json
… go/vt/vtgate/planbuilder/testdata/foreignkey_checks_on_cases.json