## Summary
This PR introduces several optimizations to improve query planning and execution performance:

1. Refactored the `Operator` interface to use `unaryOperator` and `binaryOperator` types in most cases, replacing the generic `Inputs()` and `SetInputs()` methods.
2. Implemented reuse of the `evalengine` environment to reduce allocations.
3. Delayed expression simplification to later stages of query planning.

These changes aim to reduce memory allocations, improve performance, and optimize the query planning process.

## Key Changes
1. Introduced `unaryOperator` and `binaryOperator` types and refactored most operators to use these.
2. Implemented a mechanism to reuse the `evalengine` environment across multiple operations.
3. Moved expression simplification to later stages in the query planning pipeline.
4. Updated related code to work with the new operator types and optimizations.

## Performance Impact
The changes resulted in significant performance improvements across various planner benchmarks:

- **Overall Performance**: 10.76% improvement in execution time (geometric mean)
- **Memory Allocation**: 2.74% reduction in bytes allocated per operation (geometric mean)
- **Allocation Count**: 2.13% reduction in allocations per operation (geometric mean)

Raw performance figures:

```
goos: darwin
goarch: arm64
pkg: vitess.io/vitess/go/vt/vtgate/planbuilder
cpu: Apple M1 Ultra
                                            │   before    │                after                │
                                            │   sec/op    │   sec/op     vs base                │
OLTP-20                                       223.4µ ± 0%   224.8µ ± 1%   +0.62% (p=0.001 n=10)
TPCC-20                                       2.279m ± 0%   1.661m ± 0%  -27.12% (p=0.000 n=10)
TPCH-20                                       11.33m ± 1%   10.25m ± 0%   -9.52% (p=0.000 n=10)
Planner/from_cases.json-gen4-20               7.528m ± 0%   7.382m ± 1%   -1.95% (p=0.000 n=10)
Planner/filter_cases.json-gen4-20             183.6m ± 0%   149.3m ± 0%  -18.71% (p=0.000 n=10)
Planner/large_cases.json-gen4-20              490.1µ ± 1%   470.4µ ± 0%   -4.02% (p=0.000 n=10)
Planner/aggr_cases.json-gen4-20               13.03m ± 1%   13.29m ± 0%   +2.03% (p=0.000 n=10)
Planner/select_cases.json-gen4-20             10.51m ± 1%   10.31m ± 2%   -1.88% (p=0.002 n=10)
Planner/union_cases.json-gen4-20              3.815m ± 1%   3.787m ± 0%   -0.73% (p=0.005 n=10)
SemAnalysis-20                                63.09m ± 0%   28.85m ± 1%  -54.27% (p=0.000 n=10)
SelectVsDML/DML_(random_sample,_N=32)-20      991.8µ ± 2%   978.9µ ± 2%        ~ (p=0.075 n=10)
SelectVsDML/Select_(random_sample,_N=32)-20   1.947m ± 0%   1.711m ± 0%  -12.12% (p=0.000 n=10)
BaselineVsMirrored/Baseline-20                128.6µ ± 1%   128.3µ ± 0%        ~ (p=0.143 n=10)
BaselineVsMirrored/Mirrored-20                331.4µ ± 1%   337.0µ ± 0%   +1.69% (p=0.000 n=10)
geomean                                       3.175m        2.834m       -10.76%

                                            │    before     │                 after                 │
                                            │     B/op      │     B/op       vs base                │
OLTP-20                                        145.1Ki ± 0%    145.8Ki ± 0%   +0.52% (p=0.000 n=10)
TPCC-20                                       1082.3Ki ± 0%   1017.3Ki ± 0%   -6.00% (p=0.000 n=10)
TPCH-20                                        5.328Mi ± 0%    5.230Mi ± 0%   -1.84% (p=0.000 n=10)
Planner/from_cases.json-gen4-20                4.506Mi ± 0%    4.532Mi ± 0%   +0.58% (p=0.000 n=10)
Planner/filter_cases.json-gen4-20              22.39Mi ± 0%    17.95Mi ± 0%  -19.82% (p=0.000 n=10)
Planner/large_cases.json-gen4-20               269.0Ki ± 0%    271.8Ki ± 0%   +1.03% (p=0.000 n=10)
Planner/aggr_cases.json-gen4-20                7.185Mi ± 0%    7.223Mi ± 0%   +0.53% (p=0.000 n=10)
Planner/select_cases.json-gen4-20              6.066Mi ± 0%    6.070Mi ± 0%   +0.07% (p=0.000 n=10)
Planner/union_cases.json-gen4-20               2.217Mi ± 0%    2.226Mi ± 0%   +0.41% (p=0.000 n=10)
SelectVsDML/DML_(random_sample,_N=32)-20       665.9Ki ± 0%    642.8Ki ± 0%   -3.46% (p=0.000 n=10)
SelectVsDML/Select_(random_sample,_N=32)-20    1.100Mi ± 0%    1.029Mi ± 0%   -6.46% (p=0.000 n=10)
BaselineVsMirrored/Baseline-20                 87.88Ki ± 0%    88.05Ki ± 0%   +0.19% (p=0.000 n=10)
BaselineVsMirrored/Mirrored-20                 202.8Ki ± 0%    204.5Ki ± 0%   +0.84% (p=0.000 n=10)
geomean                                        1.280Mi         1.245Mi        -2.74%

                                            │   before    │                 after                 │
                                            │  allocs/op  │  allocs/op   vs base                  │
OLTP-20                                       3.612k ± 0%   3.612k ± 0%        ~ (p=1.000 n=10) ¹
TPCC-20                                       25.70k ± 0%   24.43k ± 0%   -4.95% (p=0.000 n=10)
TPCH-20                                       134.0k ± 0%   131.6k ± 0%   -1.80% (p=0.000 n=10)
Planner/from_cases.json-gen4-20               106.7k ± 0%   106.8k ± 0%   +0.11% (p=0.000 n=10)
Planner/filter_cases.json-gen4-20             562.7k ± 0%   481.1k ± 0%  -14.51% (p=0.000 n=10)
Planner/large_cases.json-gen4-20              10.52k ± 0%   10.46k ± 0%   -0.61% (p=0.000 n=10)
Planner/aggr_cases.json-gen4-20               164.8k ± 0%   164.8k ± 0%   +0.04% (p=0.000 n=10)
Planner/select_cases.json-gen4-20             140.4k ± 0%   140.0k ± 0%   -0.27% (p=0.000 n=10)
Planner/union_cases.json-gen4-20              52.07k ± 0%   52.06k ± 0%   -0.03% (p=0.000 n=10)
SelectVsDML/DML_(random_sample,_N=32)-20      13.14k ± 0%   13.26k ± 0%   +0.89% (p=0.000 n=10)
SelectVsDML/Select_(random_sample,_N=32)-20   25.30k ± 0%   23.93k ± 0%   -5.42% (p=0.000 n=10)
BaselineVsMirrored/Baseline-20                2.276k ± 0%   2.276k ± 0%        ~ (p=1.000 n=10) ¹
BaselineVsMirrored/Mirrored-20                5.700k ± 0%   5.700k ± 0%        ~ (p=1.000 n=10) ¹
geomean                                       32.25k        31.56k        -2.13%
¹ all samples are equal
```

## Checklist
-   [x] "Backport to:" labels have been added if this change should be back-ported to release branches
-   [x] If this change is to be back-ported to previous releases, a justification is included in the PR description
-   [x] Tests were added or are not required
-   [x] Did the new or modified tests pass consistently locally and on CI?
-   [x] Documentation was added or is not required